requirements.txt
aiohttp==3.8.4
python-dateutil==2.8.2
motor==3.1.2
discord.py==2.2.3